moved components into semantic folders
