+++ src/components/FrontPageComps/Login.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import {Link, useHistory} from 'react-router-dom';
-import "../css/login.css";
+import "../../css/login.css";
 
 function Login() {
 
@@ -30,15 +30,19 @@ function Login() {
         <div className="login">
             
             <div className="login_title">
-                <p>Continuing Players log in here</p>
+                <p>Continuing Players can log in here:</p>
             </div>
 
 
             <div className = "login_inputs">
-                <p>Email:</p>
-                <input type = "text" onChange = {e => setEmail(e.target.value)}/>
-                <p>Password:</p>
-                <input type = "password" onChange = {e => setPassword(e.target.value)}/>
+                <div>
+                    <p>Email:</p>
+                    <input type = "text" onChange = {e => setEmail(e.target.value)}/>
+                </div>
+                <div>
+                    <p>Password:</p>
+                    <input type = "password" onChange = {e => setPassword(e.target.value)}/>
+                </div>
                 <button onClick = {signIn}>LOGIN</button>
             </div>
