+++ src/components/FrontPageComps/Register.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import axios from 'axios';
 import {Link, useHistory} from 'react-router-dom';
 
-import "../css/register.css";
+import "../../css/register.css";
 
 function Register() {
 
@@ -18,7 +18,7 @@ function Register() {
         
         axios.post('/auth/create', {username, email, password})
         .then(user => {
-            history.push("/rooms");
+            history.push("/landingpage");
         })
         .catch(err => {
             alert(err.response.data);
@@ -31,17 +31,24 @@ function Register() {
         <div className="register">
 
             <div className="login_title">
-                <p>New Players can create an account here</p>
+                <p>New Players can create an account here:</p>
             </div>
 
 
             <div className = "register_inputs">
-                <p>User Name:</p>
-                <input type = "text" onChange = {e => setUsername(e.target.value)}/>
-                <p>Email:</p>
-                <input type = "text" onChange = {e => setEmail(e.target.value)}/>
-                <p>Password:</p>
-                <input type = "password" onChange = {e => setPassword(e.target.value)}/>
+                <div>
+                    <p>User Name:</p>
+                    <input type = "text" onChange = {e => setUsername(e.target.value)}/>
+                </div>
+                <div>
+                    <p>Email:</p>
+                    <input type = "text" onChange = {e => setEmail(e.target.value)}/>
+                </div>
+                <div>
+                    <p>Password:</p>
+                    <input type = "password" onChange = {e => setPassword(e.target.value)}/>
+                </div>
+                
                 <button onClick = {signUp}>REGISTER</button>
             </div>
         </div>
