+++ src/css/login.css
@@ -1,15 +1,25 @@
 .login{
 
-    height: 20vh;
+    height: 40vh;
+    width: 30vw;
 
 
     border-style: solid;
     border-color: black;
     border-width: 2px;
     border-radius: 10px;
+
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-around;
 }
 
 .login_inputs {
+
+    height: 60%;
+
     display: flex;
     flex-direction: column;
     align-items: center;
