+++ src/css/register.css
@@ -1,11 +1,23 @@
 .register{
+    height: 40vh;
+    width: 30vw;
+
+
     border-style: solid;
     border-color: black;
     border-width: 2px;
     border-radius: 10px;
+
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-around;
 }
 
 .register_inputs {
+    height: 60%;
+
     display: flex;
     flex-direction: column;
     align-items: center;
